I have tried to make refactoring changes with a light touch, not changing or breaking any of the SQL at all.
Changes to the BAS statement doctype are well tested.  Changes to the PAYG Payment Summaries doctype not so much, as I did
not have enough knowledge.

Mostly I have just relocated much of the code from the __init__.py file to the classes where the functionality is being used,

Now that this refactoring is complete I know the code well enough and hope to  make some useful feature suggestions in due course.

Hope this is useful.

Alex